+++ code/core/src/main/java/org/betonquest/betonquest/quest/event/scoreboard/ScoreboardObjectiveEventFactory.java
@@ -8,13 +8,13 @@ import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.event.PlayerEvent;
 import org.betonquest.betonquest.api.quest.event.PlayerEventFactory;
 import org.betonquest.betonquest.api.quest.event.thread.PrimaryServerThreadEvent;
-
-import java.util.Locale;
+import org.betonquest.betonquest.quest.event.point.PointType;
 
 /**
  * Factory to create scoreboard events from {@link Instruction}s.
  */
 public class ScoreboardObjectiveEventFactory implements PlayerEventFactory {
+
     /**
      * Data for primary server thread access.
      */
@@ -32,25 +32,10 @@ public class ScoreboardObjectiveEventFactory implements PlayerEventFactory {
     @Override
     public PlayerEvent parsePlayer(final Instruction instruction) throws QuestException {
         final Variable<String> objective = instruction.get(Argument.STRING);
-        final String number = instruction.next();
-        final String action = instruction.getValue("action");
-        if (action != null) {
-            try {
-                final ScoreModification type = ScoreModification.valueOf(action.toUpperCase(Locale.ROOT));
-                return new PrimaryServerThreadEvent(
-                        new ScoreboardObjectiveEvent(objective, instruction.get(number, Argument.NUMBER), type),
-                        data);
-            } catch (final IllegalArgumentException e) {
-                throw new QuestException("Unknown modification action: " + instruction.current(), e);
-            }
-        }
-        if (!number.isEmpty() && number.charAt(0) == '*') {
-            return new PrimaryServerThreadEvent(
-                    new ScoreboardObjectiveEvent(objective, instruction.get(number.replace("*", ""), Argument.NUMBER), ScoreModification.MULTIPLY),
-                    data);
-        }
+        final Variable<Number> number = instruction.get(Argument.NUMBER);
+        final PointType action = instruction.getValue("action", Argument.ENUM(PointType.class), PointType.ADD).getValue(null);
         return new PrimaryServerThreadEvent(
-                new ScoreboardObjectiveEvent(objective, instruction.get(number, Argument.NUMBER), ScoreModification.ADD),
+                new ScoreboardObjectiveEvent(objective, number, action),
                 data);
     }
 }